TASK8: Add sysfs
